+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Added
+- Add option to install with homebrew
+
 ## v1.0.0 - 2023-05-23
 
 Initial release 🚀